help

+++ src/main/java/com/codeup/demo/models/Endorsement.java
@@ -17,7 +17,7 @@ public class Endorsement {
     private long id;
 
     @Column(nullable = false)
-    private byte value;
+    private int value;
 
     @Column(nullable = false)
     @JsonFormat(pattern = "mm-dd-yyyy")
@@ -41,12 +41,12 @@ public class Endorsement {
     }
 
 
-    public Endorsement(long id, byte value) {
+    public Endorsement(long id, int value) {
         this.id = id;
         this.value = value;
     }
 
-    public Endorsement(byte value, Date date, User user, Report report) {
+    public Endorsement(int value, Date date, User user, Report report) {
         this.value = value;
         this.date = date;
         this.user = user;
@@ -69,11 +69,11 @@ public class Endorsement {
         this.id = id;
     }
 
-    public byte getValue() {
+    public int getValue() {
         return value;
     }
 
-    public void setValue(byte value) {
+    public void setValue(int value) {
         this.value = value;
     }
